Broadcast changes to relations #8
This is a fairly major change to the way that the system functions. Previously we were sending an entire transaction to the socket, which included multiple changes. However creates a limitation: the user is unable to subscribe to changes to particular tables.

This change solves that. The user now receives a message for every change within a transaction. They can also subscribe to the changes like so:

realtime:* - subscribe to everything
realtime:public  - subscribe to all changes on the public schema
realtime:public:users - subscribe only to changes on the users table

This is a good step forward, but it will also be great if we can eventually allow the user to subscribe to certian rows as well. I will leave that for another day as I think it will need more thought (for example, tables with composite keys?).