Fixes #15
This is a partial implementation that allows filtering only with the "eq" filter

Users can do something like this:
realtime:public:users:id=eq.2

This will only send the notifications when the user ID 2 is updated.

There are a lot of missing filters but these will be more complicated as it will require some sort of stateful introspeciton of each room. This is a good early win that should cover many of the common cases