fix: Take version into consideration in Postgres libcluster Strategy (#678)
Currently we have nodes connecting from different versions which could have a negative impact
on the overall cluster operations. To avoid it, we're taking into consideration the version
when connecting nodes via libcluster Postgres strategy.

This can be set via a strategy option named `channel_name` and for realtime we're
using POSTGRES_CLUSTER_CHANNEL_NAME if we need to override it.