fix: Change backoff type to stop connection
DBConnection seems to not properly link to parent process.

As such, when a parent process dies, the DBConnection pool will be kept alive and keeps retrying to connect potentially creating a big issues of pending processes that keep accumulating.

This change will change the backoff strategy to :stop and let the parent process do the retrying