fix: Add write checks to channels
Add write permissions checks for the Authorization modules.

It checks if a given user can write by trying to updating a column in the channels table. After done, reverts the change within the same transaction.

Also took the time to refactor a bit the code to use some helper functions previously built in the Repo module to avoid bare sql code and have a more structured approach.