fix: Run migrations on Connect (#790)
Running migrations on Connect will ensure that the tenant database will have the required
tables for the new Realtime features. This offloads the concern from the extensions and moves
it to the Tenant domain of concerns. 

We're also call this in a sync manner to ensure that we won't fall into an issue where users can connect and have a db connection but when they check authorization they will get an error since
the migrations might not have run at that point in time.