fix: Reimplement Authz with newer simpler policies (#1029)
New Authorization checking made with simpler RLS rules. We'll now check a single table called messages and for a specific set of elements for each row.

At the moment we are prefilling the table with some information on lib/realtime/tenants/authorization.ex and that is then used to check read policies and after that we try to insert new values onto the tables to check write policies.

All the changes made during this checks is reverted upon ending the transaction ensuring we don't leave extra clutter.