fix: Shutdown Connect when migrations fail (#1151)
Currently we had an issue where Connect module was not shuting down properly leading to more and more connections to be created against the tenant database. Now we use reason :shutdown to enforce the process is killed.

Also the existing Listen/Notify approach will be deleted for a different approach to be built in the future.