fix: Delete realtime.messages entries older than 72h on connect
Starts a clean up task on each node to delete the messages in `realtime.message`. It checks:
* Which tenants need cleanup on a given region
* If the tenant should be cleaned by a given node
* Chunks the groups of tenants to avoid overloading Realtime.Repo and number of connections from a given node
* Starts connections for each tenant
* Runs query to delete messages older than 72h

We need two setups needed for tests:
* Self hosted isn't connected as a cluster nor has a region so we need a default behaviour for "regionless" setups
* For Node mode, we need to emulate the region and emulate the node connected

This impacts Cachex tests which is why we had to change to have max_cases set to 1 to avoid parallel tests from other modules to interfer

Finally, we also need to define if the application starts or not the scheduled tasks. This helps with tests but also helps if we need to stop this processes in production.