fix: syn conflict resolution (#1481)
* fix: rewrite SynHandler to pick the oldest process on conflicts

We try to keep the oldest process. If the time they were registered is exactly the same we use
their node names to decide.

The most important part is that both nodes must 100% of the time agree on the decision

Also refactored Clustered to allow for a peer to be started without
automatically connecting to the main node.

Also ensure process is killed even if exits are trapped